Add Rusteria STD lib
