--- soteria-rust/plugin-kani/std/Cargo.toml
@@ -1,14 +1,0 @@
-# Copyright Kani Contributors
-# SPDX-License-Identifier: Apache-2.0 OR MIT
-
-[package]
-# Note: this package is intentionally named std to make sure the names of
-# standard library symbols are preserved
-name = "std"
-version = "0.59.0"
-edition = "2021"
-license = "MIT OR Apache-2.0"
-publish = false
-
-[dependencies]
-kani = {path="../kani"}
