+++ soteria-rust/plugin_kani/kani/Cargo.toml
@@ -4,4 +4,5 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+rusteria = { path = "../../plugin_rusteria/rusteria" }
 kani_macros = { path = "../kani_macros" }
