+++ soteria-rust/plugin_kani/kani/src/invariant.rs
