+++ soteria-rust/plugin_rusteria/rusteria/Cargo.toml
@@ -1,0 +1,4 @@
+[package]
+name = "rusteria"
+version = "0.1.0"
+edition = "2021"
