+++ soteria-rust/plugin_rusteria/std/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "std"
+version = "0.1.0"
+edition = "2021"
+license = "MIT OR Apache-2.0"
+publish = false
+
+[dependencies]
+rusteria = {path="../rusteria"}
