+++ soteria-rust/test/cram/rusteria-lib.t/lib-fns.rs
@@ -1,0 +1,9 @@
+fn main() {
+    let cond: bool = rusteria::nondet();
+    if cond {
+        rusteria::assert(cond, "This holds");
+    } else {
+        rusteria::assume(cond);
+        rusteria::panic("Unreachable");
+    }
+}
