+++ soteria-rust/test/cram/rusteria-lib.t/run.t
@@ -1,0 +1,5 @@
+Test base functions
+  $ soteria-rust exec-main lib-fns.rs --clean
+  Done. - Ran 1 branches
+  PC: 
+    (V|0| != 0) /\ (V|0| <= 1) /\ (0 <= V|0|)
